external/ReferDINO/requirements.txt
addict==2.4.0
colorlog==6.9.0
easydict==1.13
einops==0.8.1
h5py==3.12.1
imgaug==0.4.0
ipdb==0.13.13
loralib==0.1.2
matplotlib==3.10.3
numpy==1.23.0
opencv_python==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
pycocotools==2.0.8
PyYAML==6.0.2
rich==14.0.0
ruamel.base==1.0.0
scipy==1.13.1
seaborn==0.13.2
setuptools==75.8.0
supervision==0.6.0
termcolor==3.1.0
timm==1.0.16
tqdm==4.67.1
transformers==4.48.3
wandb==0.20.1
yapf==0.43.0
scikit-image